requirements.txt
google-genai
moviepy>=2.0
Pillow
python-dotenv
requests
numpy
